+++ history.md
@@ -5,6 +5,7 @@
 - added all missing status codes
 - added parameter passing for get request using the :param key in header
 - the warning about the logger when using a string was a bad idea
+- multipart parameters names should not be escaped
 
 # 1.5.1